#837 fix more tests that broke
